Ex2/Flow/program_distribution_20240603/pyproject.toml
[tool.poetry]
name = "program-distribution-20240603"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
pillow = "^10.4.0"
matplotlib = "^3.9.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
